% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhts.R
\name{aggdhts}
\alias{aggdhts}
\title{aggregate discrete hierarchical time series}
\usage{
aggdhts(x)
}
\arguments{
\item{dhts}{a dhts object}
}
\description{
function for aggregting discrete hierarchical time series.
}
