% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{allreconcile.train}
\alias{allreconcile.train}
\title{reconcile base probabilistic forecasts}
\usage{
allreconcile.train(basef, real, lambda = 1, optimized = TRUE)
}
\arguments{
\item{basef}{joint base probabilistic forecasts}

\item{real}{dhts object holding observed series}

\item{lambda}{penalty terms of cost, default is 1.}

\item{optimized}{}
}
\value{
r*q transformation matrix
}
\description{
Method for reconciling base probabilistic forecasts
}
