% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{bu_train}
\alias{bu_train}
\title{function to train bottom-up model}
\usage{
bu_train(x)
}
\arguments{
\item{x}{dhier object}
}
\value{
reconciliation matrix
}
\description{
function to train bottom-up model
}
