% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_costeMatrix}
\alias{cal_costeMatrix}
\title{cost of moving probabilities}
\usage{
cal_costeMatrix(incoherent_domain, coherent_domain)
}
\description{
function to calculate distance between coherent point
}
