% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{coherentadj}
\alias{coherentadj}
\title{function to adjust joint distribution given new marginal distribution}
\usage{
coherentadj(x, target, hier, which)
}
\arguments{
\item{x}{original joint distribution}

\item{target}{new marginal distribution}

\item{hier}{metadata of the hierarchy}

\item{which}{integer indicating which column in the domain 
the marginal distribution refers to.}
}
\value{
adjusted joint distribution
}
\description{
function to adjust joint distribution given new marginal distribution
}
