% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhts.R
\name{construct_meta}
\alias{construct_meta}
\title{Construct metadata for a hierarchy.}
\usage{
construct_meta(domain_bts, s_mat, node_names = NULL)
}
\arguments{
\item{domain_bts}{domain of the bottom level series}

\item{s_mat}{summing matrix}
}
\description{
Construct metadata for a hierarchy.
}
