% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{construct_Q}
\alias{construct_Q}
\title{function to construct Q matrix used in optimization}
\usage{
construct_Q(r, basef, indicators = NULL)
}
\arguments{
\item{r}{cardinality of coherent domain}

\item{basef}{base forecasts}

\item{indicators}{indicators}
}
\value{
Q matrix
}
\description{
function to construct Q matrix used in optimization
}
