% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{dfr}
\alias{dfr}
\title{function to train discrete reconciliation models}
\usage{
dfr(
  hier,
  method = c("bu", "td", "dfr", "sdfr"),
  obs_train = NULL,
  bf_train = NULL,
  lambda = 0,
  optimized = TRUE
)
}
\arguments{
\item{hier}{`dhier` object representing the hierarchy.}

\item{method}{Available methods include "dfr", "sdfr", "bu", "td".}

\item{obs_train}{Past observations used for train the model.
It should be provided for all methods except "bu".}

\item{bf_train}{list containing base probabilistic forecasts of each series.
It should be provided for method "dfr" and "sdfr".}

\item{lambda}{penalty value used in "dfr"}

\item{optimized}{boolean value used in "dfr". If TRUE, the elements of 
non-nearest points in the reconciliation matrix will be zero.}
}
\value{
the reconciliation model
}
\description{
function to train discrete reconciliation models
}
