% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhier.R
\name{dhier}
\alias{dhier}
\title{discrete hierarchy constructor.}
\usage{
dhier(s_mat, domain, node_names = NULL)
}
\arguments{
\item{s_mat}{the summing matrix}

\item{domain}{the domain of bottom-level series, 2 * m matrix.
Each column corresponds to a series, and the first row corresponds to
the minimums, and the second row corresponds the maximums.}
}
\value{
a dhier object
}
\description{
Create a 'dhier' object
}
\examples{
# create a 3-nodes `dhier` object.
s_mat <- matrix(c(1, 1, 0, 1, 0, 1), 3)
domain <- matrix(c(0, 1, 0, 1), 2)
dhier(s_mat, domain)
}
