% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteRecon-package.R
\docType{package}
\name{DiscreteRecon-package}
\alias{DiscreteRecon-package}
\alias{_PACKAGE}
\title{DiscreteRecon: Probabilistic forecast reconciliation for discrete time series}
\description{
Methods and tools for reconcling probabilistic forecasts for discrete hierarchical time series.
}
\author{
\strong{Maintainer}: Bohan Zhang \email{zhangbohan@buaa.edu.cn} [copyright holder]

}
\keyword{package}
