% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{fr_train}
\alias{fr_train}
\title{reconcile base probabilistic forecasts}
\usage{
fr_train(hier, bf_train, obs_train, lambda = 0, optimized = TRUE)
}
\arguments{
\item{hier}{dhier object}

\item{bf_train}{incoherent joint base probabilistic forecasts}

\item{obs_train}{dhts object holding observed series}

\item{lambda}{penalty terms of cost, default is 0. not working if optimized=
TRUE}

\item{optimized}{indicating whether optimized. i.e., 
only moving probabilities to nearest coherent points.}
}
\value{
transformation matrix
}
\description{
Method for reconciling base probabilistic forecasts
}
