% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marginal2Joint}
\alias{marginal2Joint}
\title{convert marginal distributions into joint distribution assuming independence}
\usage{
marginal2Joint(x, domain)
}
\arguments{
\item{x}{list of distributions of all series}

\item{domain}{domain of hierarchy, if coherent, only bottom series are used (bottom-up approach).}
}
\value{
joint distribution matrix
}
\description{
convert marginal distributions into joint distribution assuming independence
}
