% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marginal2Sum}
\alias{marginal2Sum}
\title{function to compute distribution of sum of bottom series assuming independence}
\usage{
marginal2Sum(x, meta, which = 1)
}
\arguments{
\item{x}{list of distributions of some series}

\item{which}{indicating which upper nodes, default NULL means all upper series.}

\item{obj}{dhts obj}
}
\value{
distribution of upper series.
}
\description{
function to compute distribution of sum of bottom series assuming independence
}
