% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{point_forecast}
\alias{point_forecast}
\title{Point forecast}
\usage{
point_forecast(dist, meta)
}
\arguments{
\item{dist}{coherent distribution or basef}

\item{meta}{metadata}
}
\value{
point forecasts of all series
}
\description{
function to generate point forecasts of the hierarchy
}
