% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{point_metric}
\alias{point_metric}
\title{Point metrics of each series}
\usage{
point_metric(dist, y, f)
}
\arguments{
\item{y}{dhts}

\item{f}{function to calculate metric}
}
\value{
vector containing metric of each series
}
\description{
Point metrics of each series
}
