% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile.default}
\alias{reconcile.default}
\title{function to warn irregular object}
\usage{
\method{reconcile}{default}(x, ...)
}
\arguments{
\item{x}{output of reconcile.train}
}
\description{
function to warn irregular object
}
