% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile}
\alias{reconcile}
\title{generic function to reconcile base forecasts}
\usage{
reconcile(x, ...)
}
\arguments{
\item{x}{trained reconciliation model}
}
\description{
generic function to reconcile base forecasts
}
