% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile.rec_mat}
\alias{reconcile.rec_mat}
\title{regular reconciliation}
\usage{
\method{reconcile}{rec_mat}(x, basef)
}
\arguments{
\item{x}{rec_mat object}

\item{basef}{listing containing base probabilistic forecasts for each series}
}
\description{
regular reconciliation
}
