% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile.sw_res}
\alias{reconcile.sw_res}
\title{step wise reconciliation}
\usage{
\method{reconcile}{sw_res}(x, basef)
}
\arguments{
\item{x}{sw_res object}

\item{basef}{listing containing base probabilistic forecasts for each series}
}
\description{
step wise reconciliation
}
