% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile.topdown}
\alias{reconcile.topdown}
\title{topdown reconcilation function}
\usage{
\method{reconcile}{topdown}(x, basef)
}
\arguments{
\item{x}{topdown object}

\item{basef}{base forecasts of total level}
}
\description{
topdown reconcilation function
}
