% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_dfr}
\alias{reconcile_dfr}
\title{regular reconciliation}
\usage{
reconcile_dfr(mdl, bf)
}
\arguments{
\item{mdl}{trained reconciliation model}

\item{bf}{base forecasts}
}
\value{
reconciled forecasts
}
\description{
regular reconciliation
}
