% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_sdfr}
\alias{reconcile_sdfr}
\title{stepwise reconciliation}
\usage{
reconcile_sdfr(mdl, new_bf)
}
\arguments{
\item{mdl}{trained stopwise reconciliation model}

\item{new_bf}{base forecasts}
}
\value{
reconciled forecasts
}
\description{
stepwise reconciliation
}
