% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{sfr_train}
\alias{sfr_train}
\title{function to stepsise reconcile base forecasts}
\usage{
sfr_train(hier, bf_train, obs_train, lambda = 0, optimized = TRUE)
}
\arguments{
\item{hier}{dhier object}

\item{bf_train}{list containing base probabilistic forecasts of each series.
It should be provided for method "dfr" and "sdfr".}

\item{obs_train}{Past observations used for train the model}

\item{lambda}{penalty value}

\item{optimized}{boolean}
}
\value{
reconciliation matrices
}
\description{
function to stepsise reconcile base forecasts
}
