% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{Step2Joint}
\alias{Step2Joint}
\title{function to stepwise update the joint distribution and joint domain.}
\usage{
Step2Joint(ans)
}
\arguments{
\item{ans}{list containing joint distribution of all sub trees.}
}
\description{
function to stepwise update the joint distribution and joint domain.
}
