% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{stepwise_reconcile.train}
\alias{stepwise_reconcile.train}
\title{function to step-sisely reconcile base forecasts}
\usage{
stepwise_reconcile.train(basef, real, lambda = 1, optimized = TRUE)
}
\description{
function to step-sisely reconcile base forecasts
}
