% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{td_train}
\alias{td_train}
\title{function to train top-down model}
\usage{
td_train(x, obs_train)
}
\arguments{
\item{x}{dhier object}

\item{obs_train}{observations used for train}
}
\value{
reconciliation matrix
}
\description{
function to train top-down model
}
