% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{topdown.train}
\alias{topdown.train}
\title{function to train the naive top-down method based on concurrence.}
\usage{
topdown.train(real)
}
\arguments{
\item{real}{dhts object}
}
\value{
topdown object
}
\description{
function to train the naive top-down method based on concurrence.
}
