% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{twolevelhierarchy}
\alias{twolevelhierarchy}
\title{function to construct summing matrix of a simple two-level hierarchy}
\usage{
twolevelhierarchy(m)
}
\arguments{
\item{m}{number of children nodes}
}
\value{
summing matrix
}
\description{
function to construct summing matrix of a simple two-level hierarchy
}
