% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{validate_bf}
\alias{validate_bf}
\title{validator for input base forecasts}
\usage{
validate_bf(x, hier)
}
\arguments{
\item{x}{base probabilistic forecasts}

\item{hier}{hierarchy}
}
\description{
validator for input base forecasts
}
