% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{validate_obs}
\alias{validate_obs}
\title{validator for input observations}
\usage{
validate_obs(x, hier)
}
\arguments{
\item{x}{time series}

\item{hier}{hierarchy}
}
\description{
validator for input observations
}
