% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{brier_score}
\alias{brier_score}
\title{Brier Score}
\usage{
brier_score(probf, y)
}
\arguments{
\item{probf}{joint probabilistic forecasts, T * n}

\item{y}{dhts}
}
\description{
function to compute brier score given observations and probabilistic forecasts
}
