% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_costeMatrix}
\alias{cal_costeMatrix}
\title{cost of moving probabilities
function to calculate distance between coherent point}
\usage{
cal_costeMatrix(meta)
}
\arguments{
\item{meta}{metadata of dhts object}
}
\value{
distance matrix
}
\description{
cost of moving probabilities
function to calculate distance between coherent point
}
