% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhier.R
\name{cons_domain}
\alias{cons_domain}
\title{construct domain}
\usage{
cons_domain(domain, s_mat, coherent, node_names)
}
\arguments{
\item{domain}{domain matrix used in creating dhier object.}

\item{s_mat}{summing matrix}

\item{coherent}{logical indicating if produced domain is coherent or not.}
}
\value{
domain object
}
\description{
Method for constructing coherent domain vector given domain matrix of all 
bottom-level series and the summing matrix.
}
