% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cons_realDummy}
\alias{cons_realDummy}
\title{function to transform observed bottom-level series into dummy series.}
\usage{
cons_realDummy(x, hier)
}
\arguments{
\item{x}{time series}

\item{hier}{dhier object}
}
\value{
dummy matrix
}
\description{
function to transform observed bottom-level series into dummy series.
}
