% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Joint2Marginal}
\alias{Joint2Marginal}
\title{function to convert Joint distribution to marginal Distribution}
\usage{
Joint2Marginal(x, domain, which = NULL)
}
\arguments{
\item{x}{joint distribution}

\item{domain}{domain of hierarchy, incoherent or coherent}

\item{which}{integer indicating which dimension (which column of domain), if
NULL, return marginal distribution of all series.}
}
\value{
marginal distribution
}
\description{
function to convert Joint distribution to marginal Distribution
}
