% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marginal2Joint}
\alias{marginal2Joint}
\title{method for incoherent probabilistic forecasts}
\usage{
marginal2Joint(x, domain)
}
\description{
method for incoherent probabilistic forecasts
}
\note{
This method assume the list containing base forecasts is ordered by the 
order of summing matrix and probabilities are arranged in order of domain.
}
