% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marginal2Joint}
\alias{marginal2Joint}
\title{convert marginal distributions into joint distribution assuming independence}
\usage{
marginal2Joint(x, meta, method)
}
\arguments{
\item{x}{list of distributions of all series}

\item{method}{bu for bottom-up, produce coherent joint distribution; 
ind for independent, producing incoherent base joint distribution.}

\item{obj}{dhts object}
}
\value{
joint distribution matrix
}
\description{
convert marginal distributions into joint distribution assuming independence
}
