% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{stepwise_reconcile.train}
\alias{stepwise_reconcile.train}
\title{function to step-sisely reconcile base forecasts}
\usage{
stepwise_reconcile.train(basef, real, lambda = 0, optimized = TRUE)
}
\arguments{
\item{basef}{basef forecasts}

\item{real}{dhts}

\item{lambda}{penalty}

\item{optimized}{if optimization}
}
\value{
sw_res object holding reconciliation matrix of each step
}
\description{
function to step-sisely reconcile base forecasts
}
